Added basic combat and hopefully fixed jumping

+++ Assets/Scripts/AttackPattern.cs
@@ -4,6 +4,6 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "AttackPattern", menuName = "Scriptable Objects/AttackPattern")]
 public class AttackPattern : ScriptableObject
 {
-    public float damage;
-    
+    public int damage;
+
 }
